requirements.txt
google-genai 
sentence-transformers 
pinecone 
py2neo
python-dotenv
numpy